+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Version 5.0.3 (In development)
 
+### App Center Distribute
+
+* **[Fix]** Add RECEIVER_EXPORTED flag for install receiver.
+* **[Fix]** Add FLAG_ALLOW_UNSAFE_IMPLICIT_INTENT flag for broadcast pending intent.
+
 ## Version 5.0.2
 
 ### AppCenter